T#773: Extract daemons (notify drain + DB maintenance + file archive) to src/daemons/routes.ts
Phase 1.8 of Backend Modularization (Library #102). Mechanical move, no
logic changes.

**Moved to src/daemons/routes.ts (~440L)**:
- Notification Queue Drain (Spec #54): perBeastDrainAlive() + runDrainCycle()
  + DRAIN_INTERVAL/DRAIN_SPACING/DRAIN_DIR/drainLastSent state. PID-reuse
  defense via /proc/<pid>/cmdline check preserved verbatim.
- DB Maintenance: runDbMaintenance() + DB_RETENTION_DAYS/DB_MAINTENANCE_INTERVAL.
  Audit log retention, security event pruning (90d), beast token pruning (7d
  post-expiry), VACUUM reclamation.
- File Archive (T#533): runFileArchive() + FILE_ARCHIVE_GRACE_DAYS/INTERVAL.
  Soft-deleted files → compressed tar.gz archives after 7d grace.
- 5 daemon-management routes:
  POST /api/db/maintenance, GET /api/db/stats
  GET /api/files/archive/stats, POST /api/files/archive/run, POST /api/files/:id/restore

**Exports**:
- registerDaemonRoutes(app, sqlite) — registers the 5 management routes
  with closure-shadowed sqlite (non-null param)
- initDaemons(sqliteDb) — captures sqlite to module state, starts all 3
  daemon setInterval/setTimeout cycles once via daemonsStarted guard

**Imports from**:
- ../server/security-logger.ts: SECURITY_RETENTION_DAYS, pruneSecurityEvents
- ../server/beast-tokens.ts: pruneBeastTokens
- ../config.ts: ORACLE_DATA_DIR (UPLOADS_DIR + ARCHIVE_DIR derived locally
  via path.join — same shape as the inline derivations in server.ts)

**Type safety pattern**:
- Module-level moduleSqlite (Database | null) shadowed inside each daemon
  function via 'if (!moduleSqlite) return; const sqlite = moduleSqlite;'
  null-narrow guard (defensible per Bertus T#772 review — daemons must not
  throw or they'd crash the timer loop; init-order is synchronous before
  any daemon could fire)
- registerDaemonRoutes uses 'const sqlite = sqliteDb' closure (always non-null)

**Server.ts changes**:
- Added import (next to initScheduler/registerScheduler imports)
- Removed L6323-6706 (drain + maint + file blocks, ~384 lines) — replaced
  with 4-line init+register call block
- No other server.ts changes (Withings auto-sync left in place for Phase
  1.10 integrations extraction — it depends on syncWithingsMeasurements
  which is integration-specific)

**Verification**:
- Boot test: server starts clean, three daemon log lines printed exactly
  once: '[Notify] Queue drain started', '[DB Maintenance] Retention: 15
  days', '[File Archive] Grace: 7 days'
- tsc --noEmit: zero new errors in src/daemons/routes.ts
- No logic changes: 1:1 mechanical move with closure-based type-safety
  shims. All 5 routes preserved including the POST /api/db/maintenance
  query-as 403-block, file restore 404+400 guards, etc.
- Net: server.ts down ~378 lines (was 10,861, now 10,483)
